add kmeans clustering with c implementation

+++ rust_poker_ai/src/distance.rs
@@ -2,11 +2,11 @@ use emd::distance;
 use ndarray::Array1;
 
 
-pub fn wasserstein(v1: &Vec<f32>, v2: &Vec<f32>) -> f32 {
+pub fn wasserstein(v1: &Vec<f64>, v2: &Vec<f64>) -> f64 {
     // Ensure both vectors have the same length
     assert_eq!(v1.len(), v2.len(), "Input vectors must have the same length");
 
     let arr1: Array1<f64> = v1.iter().map(|&x| x.into()).collect::<Array1<f64>>();
     let arr2: Array1<f64> = v2.iter().map(|&x| x.into()).collect::<Array1<f64>>();
-    distance(&arr1.view(), &arr2.view()) as f32
+    distance(&arr1.view(), &arr2.view()) as f64
 }
